pyproject.toml
[project]
name = "uav-rl-navigation"
version = "0.1.0"
description = "A Reinforcement Learning-based UAV navigation system for obstacle avoidance in dynamic environments."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.2.2",
    "ale-py==0.10.2",
    "annotated-types==0.7.0",
    "appnope==0.1.4",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "binaryornot==0.4.4",
    "box2d-py==2.3.5",
    "certifi==2025.1.31",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cookiecutter==2.6.0",
    "cookiecutter-data-science==2.1.0",
    "copier==9.6.0",
    "cycler==0.12.1",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "dunamai==1.23.1",
    "executing==2.2.0",
    "Farama-Notifications==0.0.4",
    "filelock==3.18.0",
    "fonttools==4.56.0",
    "fsspec==2025.3.0",
    "funcy==2.0",
    "ghp-import==2.1.0",
    "grpcio==1.71.0",
    "gymnasium==1.1.1",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "ipykernel==6.29.5",
    "ipython==9.0.2",
    "ipython_pygments_lexers==1.1.1",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "jinja2-ansible-filters==1.3.2",
    "jupyter_client==8.6.3",
    "jupyter_core==5.7.2",
    "kiwisolver==1.4.8",
    "llvmlite==0.44.0",
    "loguru==0.7.3",
    "Markdown==3.7",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mergedeep==1.3.4",
    "mkdocs==1.6.1",
    "mkdocs-get-deps==0.2.0",
    "moviepy==2.1.2",
    "mpmath==1.3.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numba==0.61.0",
    "numpy==1.26.4",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "pettingzoo==1.24.3",
    "pexpect==4.9.0",
    "pillow==10.4.0",
    "platformdirs==4.3.7",
    "plumbum==1.9.0",
    "proglog==0.1.11",
    "prompt_toolkit==3.0.50",
    "protobuf==6.30.2",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pybullet==3.2.7",
    "pydantic==2.11.3",
    "pydantic_core==2.33.1",
    "PyFlyt==0.29.0",
    "pygame==2.6.1",
    "Pygments==2.19.1",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-slugify==8.0.4",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "pyyaml_env_tag==0.1",
    "pyzmq==26.3.0",
    "questionary==2.1.0",
    "requests==2.32.3",
    "rich==13.9.4",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "setuptools==77.0.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "stable_baselines3==2.6.0",
    "stack-data==0.6.3",
    "swig==4.3.0",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "text-unidecode==1.3",
    "torch==2.6.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.15.2",
    "types-python-dateutil==2.9.0.20241206",
    "typing-inspection==0.4.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "Werkzeug==3.1.3",
    "wheel==0.45.1",
]

[dependency-groups]
dev = [
    "uav-rl-navigation",
]

[tool.uv.sources]
uav-rl-navigation = { workspace = true }

